Arel::Nodes::BitString
